Adding cmake cache variables and helper script (#16)
* also adding ./fortran in the ignore list

* Adding an alternative fix for the case sensitive file names

For some reason the auto fix mechanism in the `CMakeLists.txt` file
didn't work for me. Sometimes it never fixed the included and gfortran
complained that `common.for` is not found. Sometimes
`${CMAKE_CURRENT_BINARY_DIR}/FORTRAN_temp` was created and the fix
worked, gfortran did not complain anymore, but the linking failed with:

    Writing the .pyf file with numpy.f2py, please be patient...
    Traceback (most recent call last):
      File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 172, in <module>
        dd = generate_function_dict('REFPROP.pyf')
      File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 131, in generate_function_dict
        output = find_subroutine(lines, istart)
      File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 91, in find_subroutine
        raise ValueError(lines[istart + 1 + offset].strip())
    ValueError: real dimension((ncmax)) :: x0
    CMakeFiles/refprop.dir/build.make:519: recipe for target 'librefprop.so' failed

So I wrote a small bash script that creates lower cased symlinks of
`COMMON.FOR` and `COMTRN.FOR`.

* Adding cmake cache variable REFPROP_SOURCE_AUTOFIX

The auto fix mechanism doesn't work properly for. In most cases
compiling the fortran code was not the issue, as
`${CMAKE_CURRENT_BINARY_DIR}/FORTRAN_temp` was used for the source.
Linking was the issue:

        Writing the .pyf file with numpy.f2py, please be patient...
        Traceback (most recent call last):
          File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 172, in <module>
            dd = generate_function_dict('REFPROP.pyf')
          File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 131, in generate_function_dict
            output = find_subroutine(lines, istart)
          File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 91, in find_subroutine
            raise ValueError(lines[istart + 1 + offset].strip())
        ValueError: real dimension((ncmax)) :: x0
        CMakeFiles/refprop.dir/build.make:519: recipe for target 'librefprop.so' failed

I don't understand why this happend and I didn't want to go through
another repo fixing things for me.

So the easisest way was to add `REFPROP_SOURCE_AUTOFIX`, which defaults to
TRUE.

With  `cmake -DREFPROP_SOURCE_AUTOFIX=OFF` the whole autofix mechanism
is disabled and you can fix the file name problem yourself, for example
using the `utils/alternative_case_sens_fix` script from this repository.

* adding cache variable REFPROP_ADDITIONAL_LINK_FLAGS

I have to compile this library with 32bit support. On a debian 8 I had
to do

     $ apt-get install libgfortran-4.9-dev:i386

Which installs the 32bit gfortran library in
`/usr/lib/gcc/i586-linux-gnu/4.9

    /usr/bin/ld: skipping incompatible /usr/lib/gcc/x86_64-linux-gnu/4.9/libgfortran.so when searching for -lgfortran
    /usr/bin/ld: skipping incompatible /usr/lib/gcc/x86_64-linux-gnu/4.9/libgfortran.a when searching for -lgfortran
    /usr/bin/ld: cannot find -lgfortran
    collect2: error: ld returned 1 exit status
    CMakeFiles/refprop.dir/build.make:519: recipe for target 'librefprop.so' failed

Adding `-L /usr/lib/gcc/i586-linux-gnu/4.9` helped. However appending a
value in `REFPROP_LINK_FLAGS` from the command line is not possible
beacuse this variable is not a cache variable, so using
`-DREFPROP_LINK_FLAGS` did not work.

`REFPROP_ADDITIONAL_LINK_FLAGS` allows you to pass additional linker
options.

* Revert "Adding an alternative fix for the case sensitive file names"

This reverts commit 4c558766506ab15baba71ba05fc0ea9a6763b621.